+++ rt_thread_3.1.3-ucosiii_3.03-wrapper/uCOS-III_Wrapper/uCOS-III/os_sem.c
@@ -239,15 +239,18 @@ OS_OBJ_QTY  OSSemDel (OS_SEM  *p_sem,
     switch (opt)
     {
         case OS_OPT_DEL_NO_PEND:
+            CPU_CRITICAL_ENTER();
             if(rt_list_isempty(&(p_sem->Sem.parent.suspend_thread)))/*若没有线程等待信号量*/
             {
+                CPU_CRITICAL_EXIT();
                 rt_err = rt_sem_detach(&p_sem->Sem);
                 *p_err = _err_rtt_to_ucosiii(rt_err);                 
             }
             else
             {
+                CPU_CRITICAL_EXIT();
                 *p_err = OS_ERR_TASK_WAITING;
-            }
+            } 
             break;
             
         case OS_OPT_DEL_ALWAYS:
@@ -256,9 +259,9 @@ OS_OBJ_QTY  OSSemDel (OS_SEM  *p_sem,
             break;
     }
     
-    OS_CRITICAL_ENTER();
+    CPU_CRITICAL_ENTER();
     pend_sem_len = rt_list_len(&(p_sem->Sem.parent.suspend_thread));
-    OS_CRITICAL_EXIT();
+    CPU_CRITICAL_EXIT();
     
     return pend_sem_len;
 }
@@ -437,10 +440,136 @@ OS_SEM_CTR  OSSemPend (OS_SEM   *p_sem,
 */
 
 #if OS_CFG_SEM_PEND_ABORT_EN > 0u
+
+/*由rt_ipc_list_resume函数改编*/
+rt_inline rt_err_t rt_ipc_pend_abort_1 (rt_list_t *list)
+{
+    struct rt_thread *thread;
+    
+    CPU_SR_ALLOC();
+    
+    /* get thread entry */
+    CPU_CRITICAL_ENTER();
+    thread = rt_list_entry(list->next, struct rt_thread, tlist);
+    CPU_CRITICAL_EXIT();
+    
+    /* resume it */
+    rt_thread_resume(thread);
+
+    return RT_EOK;
+}
+
+/*由rt_ipc_list_resume_all函数改编*/
+rt_inline rt_err_t rt_ipc_pend_abort_all (rt_list_t *list)
+{
+    struct rt_thread *thread;
+
+    CPU_SR_ALLOC();
+
+    /* wakeup all suspend threads */
+    while (!rt_list_isempty(list))
+    {
+        /* disable interrupt */
+        CPU_CRITICAL_ENTER();
+
+        /* get next suspend thread */
+        thread = rt_list_entry(list->next, struct rt_thread, tlist);
+
+        /*
+         * resume thread
+         * In rt_thread_resume function, it will remove current thread from
+         * suspend list
+         */
+        rt_thread_resume(thread);
+
+        /* enable interrupt */
+        CPU_CRITICAL_EXIT();
+    }
+
+    return RT_EOK;
+}
+
 OS_OBJ_QTY  OSSemPendAbort (OS_SEM  *p_sem,
                             OS_OPT   opt,
                             OS_ERR  *p_err)
 {
+    rt_uint32_t pend_sem_len;
+    
+    CPU_SR_ALLOC();
+
+#ifdef OS_SAFETY_CRITICAL
+    if (p_err == (OS_ERR *)0) {
+        OS_SAFETY_CRITICAL_EXCEPTION();
+        return ((OS_OBJ_QTY)0u);
+    }
+#endif
+
+#if OS_CFG_CALLED_FROM_ISR_CHK_EN > 0u
+    if (OSIntNestingCtr > (OS_NESTING_CTR)0u) {             /* Not allowed to Pend Abort from an ISR                  */
+       *p_err =  OS_ERR_PEND_ABORT_ISR;
+        return ((OS_OBJ_QTY)0u);
+    }
+#endif
+
+#if OS_CFG_ARG_CHK_EN > 0u
+    if (p_sem == (OS_SEM *)0) {                             /* Validate 'p_sem'                                       */
+       *p_err =  OS_ERR_OBJ_PTR_NULL;
+        return ((OS_OBJ_QTY)0u);
+    }
+    switch (opt) {                                          /* Validate 'opt'                                         */
+        case OS_OPT_PEND_ABORT_1:
+        case OS_OPT_PEND_ABORT_ALL:
+        case OS_OPT_PEND_ABORT_1   | OS_OPT_POST_NO_SCHED:
+        case OS_OPT_PEND_ABORT_ALL | OS_OPT_POST_NO_SCHED:
+             break;
+
+        default:
+            *p_err =  OS_ERR_OPT_INVALID;
+             return ((OS_OBJ_QTY)0u);
+    }
+#endif
+    
+#if OS_CFG_OBJ_TYPE_CHK_EN > 0u
+    if (rt_object_get_type(&p_sem->Sem.parent.parent) != RT_Object_Class_Semaphore) {/*Make sure semaphore was created*/
+       *p_err =  OS_ERR_OBJ_TYPE;
+        return ((OS_OBJ_QTY)0u);
+    }
+#endif  
+    
+    CPU_CRITICAL_ENTER();
+    if(rt_list_isempty(&(p_sem->Sem.parent.suspend_thread)))/*若没有线程等待信号量*/ 
+    {
+        CPU_CRITICAL_EXIT(); 
+       *p_err =  OS_ERR_PEND_ABORT_NONE;
+        return ((OS_OBJ_QTY)0u);        
+    }
+    else
+    {
+        CPU_CRITICAL_EXIT();
+    }
+    
+    if(opt&OS_OPT_PEND_ABORT_ALL)
+    {
+        /*OS_OPT_PEND_ABORT_ALL*/
+        rt_ipc_pend_abort_all(&(p_sem->Sem.parent.suspend_thread));
+    }
+    else
+    {
+        /*OS_OPT_PEND_ABORT_1*/
+        rt_ipc_pend_abort_1(&(p_sem->Sem.parent.suspend_thread));
+    }
+    
+    if(!opt&OS_OPT_POST_NO_SCHED)
+    {
+        rt_schedule();
+    }
+    
+    *p_err = OS_ERR_NONE;
+    CPU_CRITICAL_ENTER();
+    pend_sem_len = rt_list_len(&(p_sem->Sem.parent.suspend_thread));
+    CPU_CRITICAL_EXIT();
+    
+    return pend_sem_len;
 }
 #endif