+++ rt_thread_3.1.3-ucosiii_3.03-wrapper/changelog.md
@@ -75,6 +75,13 @@
 - 修复OSQDel函数无法返回当前等待信号量数量、无法使用OS_OPT_DEL_NO_PEND选项的问题
 - 修复OSFlagDel函数无法返回当前等待信号量数量、无法使用OS_OPT_DEL_NO_PEND选项的问题
 
+### 2020-7-10
+
+- 增加OS_TCB结构体.StkLimitPtr .StkBasePtr .PendStatus成员变量
+- 实现OSSemPendAbort函数
+
+
+
 
 
 
@@ -110,7 +117,9 @@
 
 **[bug fix]** 修复OSFlagDel函数无法返回当前等待信号量数量、无法使用OS_OPT_DEL_NO_PEND选项的问题（已知）
 
-**[add]** 实现OSTimeDlyResume函数（目前仅有12个uCOS-III API无法兼容）；
+**[add]** 实现OSTimeDlyResume函数（目前仅有12个uCOS-III API无法兼容）
+
+**[add]** 实现OSSemPendAbort函数（目前仅有11个uCOS-III API无法兼容）
 
 
 
@@ -118,3 +127,4 @@
 
 - [ ] 继续完善统计任务的其他统计项
 - [ ] 继续努力实现受限项opts
+- [ ] 继续努力实现剩余未兼容的API